+++ src/main/scala-2.11/markets/tradables/orders/Preferences.scala
@@ -13,7 +13,9 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.tradables
+package markets.tradables.orders
+
+import markets.tradables.Tradable
 
 
 /** A mixin trait that uses a total `Ordering` to express preferences over a particular type of `Tradable`.
@@ -22,7 +24,7 @@ package markets.tradables
   * @note any `Ordering` implies a `max` operator that can be used as an `operator` to compare two `Tradable` instances.
   */
 trait Preferences[T <: Tradable] extends Operator[T] {
-  this: Tradable =>
+  this: Order =>
 
   /** An `Ordering` defined over a particular type of `Tradable`. */
   def ordering: Ordering[T]
