+++ src/main/scala-2.11/markets/tradables/orders/limit/LimitAskOrder.scala
@@ -17,8 +17,8 @@ package markets.tradables.orders.limit
 
 import java.util.UUID
 
-import markets.tradables.orders.{AskOrder, BidOrder}
-import markets.tradables.{Predicate, Tradable}
+import markets.tradables.orders.{AskOrder, BidOrder, Predicate}
+import markets.tradables.Tradable
 
 
 /** Class representing an order to sell some Tradable at some price.
